requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
gunicorn
sqlalchemy==2.0.43
jinja2
python-multipart
itsdangerous
bcrypt==5.0.0
pillow